Consolidate main loop and abstract the stuff into game_mode.
game_mode is a class which can process events, respond to animation ticks, and
render stuff. To actually make a game, we'll just make a new subclass that is
the game and does those things. Now all of the generic event handling stuff gets
processed in one place in main instead of in two separate event loops.